rename `vm2` -> `vm`
