+++ src/vm/tests.rs
@@ -1,46 +1,155 @@
-use std::cell::{Cell, RefCell};
-use std::fmt::Write;
-use std::panic::catch_unwind;
+use std::sync::Arc;
 
-use super::*;
-use crate::code::print::DisplayModule;
-use crate::util::JoinIter as _;
+use super::operands::FunctionId;
+use super::value::Value;
+use super::{asm, dispatch, Error, Function, Module, Vm};
 
-thread_local! {
-    static EVENTS: RefCell<Vec<String>> = const { RefCell::new(Vec::new()) };
-    static COLLECT_EVENTS: Cell<bool> = const { Cell::new(true) };
+fn init_vm() -> Vm {
+    // always test with starting stack size of `1` to
+    // properly exercise the stack growth logic
+    Vm::builder().initial_stack_size(1).build()
 }
 
-fn debug_hook(event: DebugEvent) {
-    if !COLLECT_EVENTS.get() {
-        return;
-    }
-
-    const ALIGN: usize = 24;
-    EVENTS.with_borrow_mut(|events| match event {
-        DebugEvent::Dispatch { op, stack_frame } => {
-            let op = op.to_string();
-            events.push(format!("{op:ALIGN$} [{stack_frame}]"));
-        }
-        DebugEvent::PostDispatch { stack_frame } => {
-            let prev = events.last_mut().unwrap();
-            write!(prev, " -> [{stack_frame}]").unwrap();
-        }
-        DebugEvent::Call { call_frame } => events.push(format!("{}", call_frame)),
+fn init_module(entry: FunctionId, functions: Vec<Function>) -> Arc<Module> {
+    Arc::new(Module {
+        entry,
+        functions,
+        external_functions: vec![],
     })
 }
 
-fn get_panic_message(payload: Box<dyn std::any::Any + Send>) -> String {
-    match payload.downcast::<&'static str>() {
-        Ok(s) => String::from(*s),
-        Err(payload) => match payload.downcast::<String>() {
-            Ok(s) => *s,
-            Err(_) => "<failed to get panic message>".into(),
-        },
-    }
+macro_rules! asm_function {
+    ($name:ident ($registers:expr) {
+        $($inst:ident $($operand:expr)*);*
+        $(;)?
+    }) => (Function::new(
+        stringify!($name),
+        vec![
+            $(asm::$inst($($operand),*),)*
+        ],
+        vec![],
+        $registers,
+    ));
 }
 
-fn _run(input: &str) -> String {
+#[test]
+fn manual_simple_arithmetic() {
+    let m = init_module(
+        0,
+        vec![asm_function![add_i64(2) {
+            load_i16 0 10;
+            load_i16 1 10;
+            add_i64 0 1 0;
+            ret;
+        }]],
+    );
+
+    let mut vm = init_vm();
+    dispatch(&mut vm, &m).unwrap();
+
+    assert!(matches!(vm.ret(), Value::I64(20)), "{:?}", vm.ret());
+    assert_eq!(vm.stack_size(), 2);
+}
+
+#[test]
+fn manual_recursive_fibonacci() {
+    let r0 = 0u8;
+    let r1 = 1u8;
+    let r2 = 2u8;
+    let r3 = 3u8;
+    let r4 = 4u8;
+    let f1 = 1u16;
+
+    // fib(3) + fib(7)
+
+    let m = init_module(
+        0,
+        vec![
+            asm_function![main(3) {
+                load_i16 r1 3i16;
+                call r0 f1;
+                load_i16 r2 7i16;
+                call r1 f1;
+                add_i64 r0 r0 r1;
+                ret;
+            }],
+            asm_function![fib(5) {
+                load_i16 r2 2i16;
+                cmp_lt_i64 r0 r1 r2;
+                jump_if_false r0 3;
+                mov r0 r1;
+                jump 9;
+                load_i16 r4 1i16;
+                sub_i64 r3 r1 r4;
+                call r2 f1;
+                mov r0 r2;
+                load_i16 r4 2i16;
+                sub_i64 r3 r1 r4;
+                call r2 f1;
+                add_i64 r0 r0 r2;
+                ret;
+            }],
+        ],
+    );
+
+    let mut vm = init_vm();
+    dispatch(&mut vm, &m).unwrap();
+
+    assert_eq!(
+        vm.ret().i64().unwrap(),
+        iter_fib(3) + iter_fib(7),
+        "{:?}",
+        vm.ret()
+    );
+}
+
+fn iter_fib(n: i64) -> i64 {
+    let mut a = 0;
+    let mut b = 1;
+    for _ in 0..n {
+        let c = a + b;
+        a = b;
+        b = c;
+    }
+
+    a
+}
+
+#[test]
+fn invalid_op() {
+    let m = init_module(0, vec![asm_function![invalid_op(0) { invalid_op; }]]);
+
+    let mut vm = init_vm();
+    assert!(matches!(dispatch(&mut vm, &m), Err(Error::InvalidOp)));
+}
+
+// Compile-test:
+#[test]
+fn extern_() {
+    use super::{function, Scope};
+
+    fn _unit(_: Scope) {}
+    fn _int(_: Scope, _: i64) -> i64 {
+        0
+    }
+    fn _int2(_: Scope, _: i64, _: i64) -> i64 {
+        0
+    }
+    fn _num(_: Scope, _: f64) -> f64 {
+        0.0
+    }
+    fn _bool(_: Scope, _: bool) -> bool {
+        false
+    }
+
+    function!(_unit);
+    function!(_int);
+    function!(_int2);
+    function!(_num);
+    function!(_bool);
+}
+
+fn _emit(input: &str) -> String {
     let ast = match crate::syn::try_parse(input) {
         Ok(ast) => ast,
         Err(e) => panic!("{e}"),
@@ -49,45 +158,22 @@ fn _run(input: &str) -> String {
         Ok(hir) => hir,
         Err(e) => panic!("{e}"),
     };
-    let code = match crate::code::compile(hir) {
-        Ok(code) => code,
+    let library = crate::Library::new();
+    let module = match crate::code::compile(hir, &library) {
+        Ok(m) => m,
         Err(e) => panic!("{e}"),
     };
 
-    let mut out = String::new();
-    writeln!(
-        &mut out,
-        "## Module\n\n{}",
-        DisplayModule(&code, Some(input))
-    )
-    .unwrap();
-
-    write!(&mut out, "## Output\n\n").unwrap();
-    let module = code.link().unwrap();
-    match catch_unwind(|| Vm::new().run_debug(&module, debug_hook)) {
-        Ok(Ok(value)) => {
-            writeln!(&mut out, "{value:?}\n").unwrap();
-        }
-        Ok(Err(e)) => {
-            writeln!(&mut out, "{e}\n").unwrap();
-        }
-        Err(payload) => {
-            writeln!(&mut out, "PANIC: {}\n", get_panic_message(payload)).unwrap();
-        }
-    };
-
-    if COLLECT_EVENTS.get() {
-        EVENTS.with_borrow(|events| {
-            writeln!(&mut out, "## Events\n\n{}", events.iter().join("\n")).unwrap();
-        });
+    let mut vm = crate::Vm::new();
+    match vm.run(&module) {
+        Ok(value) => format!("{value:?}"),
+        Err(e) => format!("error: {e}"),
     }
-
-    out
 }
 
-macro_rules! run {
+macro_rules! emit {
     ($input:literal) => {
-        _run(indoc::indoc!($input))
+        _emit(indoc::indoc!($input))
     };
 }
 
@@ -95,15 +181,7 @@ macro_rules! test {
     ($name:ident, $input:literal) => {
         #[test]
         fn $name() {
-            assert_snapshot!(run!($input))
-        }
-    };
-    ($name:ident, silent, $input:literal) => {
-        #[test]
-        fn $name() {
-            COLLECT_EVENTS.set(false);
-            assert_snapshot!(run!($input));
-            COLLECT_EVENTS.set(true);
+            assert_snapshot!(emit!($input))
         }
     };
 }
@@ -178,14 +256,13 @@ test! {
 
 test! {
     fibonacci,
-    silent,
     r#"
         fn fib(n: int) -> int {
             if n < 2 { n }
             else { fib(n - 1) + fib(n - 2) }
         }
 
-        fib(3) + fib(7)
+        fib(25)
     "#
 }
